\name{read.cep}
\alias{read.cep}

\title{Reads a CEP (Canoco) data file }
\description{
  \code{read.cep} reads a file formatted by relaxed strict CEP format
  used by \code{Canoco} software, among others. This function has been
  removed from \pkg{vegan}, and it is now available in separate
  package \pkg{cepreader} as function \code{\link[cepreader]{readCEP}}.
}

\keyword{ IO }
\keyword{ file }




