\name{sipoo}
\alias{sipoo}
\docType{data}
\title{ Birds in the Archipelago of Sipoo (Sibbo)}
\description{
Land birds on islands covered by 
coniferous forest in the Sipoo archipelago, southern Finland (landbridge/
oceanic distinction unclear from source). 
}
\usage{data(sipoo)}
\format{
  A data frame with 18 sites and 50 species.
}
\source{
\url{http://www.aics-research.com/nested/}
}
\references{
Simberloff, D. & Martin, J.-L.  (1991). 
Nestedness of insular avifaunas:
simple summary statistics masking
complex species patterns.  \emph{Ornis
Fennica} 68:178--192. 
}
\examples{
data(sipoo)
}
\keyword{datasets}
