\name{dune.taxon}
\alias{dune.taxon}
\docType{data}
\title{Taxonomic Classification of Dune Meadow Species}
\description{
  Classification table of the species in the \code{\link{dune}} data
  set.
}
\usage{data(dune.taxon)}
\format{
  A data frame with 30 species (rows) classified into five taxonomic
  levels (columns).
}
\details{
  The classification of vascular plants is adapted from AGP (2003), and
  that of mosses from Hill et al. (2006).
}
\references{
  AGP [Angiosperm Phylogeny Group] (2003) An update of the Angiosperm
  Phylogeny Group classification for the orders and families of flowering
  plants: AGP II. \emph{Bot. J. Linnean Soc.} \strong{141}: 399--436.

  Hill, M.O et al. (2006) An annotated checklist of the mosses of Europe
  and Macaronesia. \emph{J. Bryology} \strong{28}: 198--267.
}
\note{
  The data set was made to demonstrate \code{\link{taxondive}}, and will
  probably be removed after a better example is found.
}
\examples{
data(dune.taxon)
}
\keyword{datasets}
