\encoding{UTF-8}
\name{vegan-deprecated}

\alias{vegan-deprecated}
\alias{humpfit}

%------ NOTE:  ../R/vegan-deprecated.R   must be synchronized with this!
\title{Deprecated Functions in vegan package}
%------ PLEASE: one \alias{.} for EACH ! (+ one \usage{} & \arguments{} for all)

\description{
  These functions are provided for compatibility with older versions of
  \pkg{vegan} only, and may be defunct as soon as the next release.
}

\usage{
## no deprecated functions at the moment
}

\details{

  Function \code{humpfit} was moved to the \pkg{natto} package and is
  still available from \url{https://github.com/jarioksa/natto}. The
  function fits a neutral no-interaction model for the biomass --
  species-richness relation. The old help page is still available with
  \code{help("humpfit-deprecated")} (note the quotes).

}

\seealso{
  \code{\link{Deprecated}}
}
\keyword{misc}
