\encoding{UTF-8}
\name{vegan-deprecated}

\alias{vegan-deprecated}
%------ NOTE:  ../R/vegan-deprecated.R   must be synchronized with this!
\title{Deprecated Functions in vegan package}
%------ PLEASE: one \alias{.} for EACH ! (+ one \usage{} & \arguments{} for all)
\description{
  These functions are provided for compatibility with older versions of
  \pkg{vegan} only, and may be defunct as soon as the next release.
}

\seealso{
  \code{\link{Deprecated}}
}
\keyword{misc}
