\name{vegandocs}
\alias{vegandocs}

\title{Display vegan Documentation}

\description{Deprecated function to display news and vignettes in
  \pkg{vegan}. You should use \code{\link[utils]{news}} or
  \code{\link[utils]{browseVignettes}} instead. }

\usage{
vegandocs(doc = c("NEWS", "ONEWS", "FAQ-vegan", "intro-vegan",
    "diversity-vegan", "decision-vegan", "partitioning", "permutations"))
}

\arguments{
  \item{doc}{The name of the document (partial match, case sensitive). }
}

\author{Jari Oksanen }

\note{The function is deprecated.}

\seealso{\code{\link[utils]{browseVignettes}},
\code{\link[utils]{vignette}}, \code{\link[utils]{news}}. }

\keyword{ utilities }
\keyword{ documentation }
